+++ Specs/DataSources/CzmlDataSourceSpec.js
@@ -3278,6 +3278,31 @@ defineSuite([
         expect(entity.polyline.material.color.getValue()).toEqual(new Color(0.1, 0.2, 0.3, 0.4));
     });
 
+    it('Polyline dash.', function() {
+        var packet = {
+            id : 'polylineDash',
+            polyline : {
+                material : {
+                    polylineDash : {
+                        color : {
+                            rgbaf : [0.1, 0.2, 0.3, 0.4]
+                        },
+                        dashLength: 16.0,
+                        dashPattern: 7.0
+                    }
+                }
+            }
+        };
+
+        var dataSource = new CzmlDataSource();
+        dataSource.load(makePacket(packet));
+
+        var entity = dataSource.entities.getById('polylineDash');
+        expect(entity.polyline.material.color.getValue()).toEqual(new Color(0.1, 0.2, 0.3, 0.4));
+        expect(entity.polyline.material.dashLength.getValue()).toEqual(16.0);
+        expect(entity.polyline.material.dashPattern.getValue()).toEqual(7.0);
+    });
+
     it('Processes extrapolation options', function() {
         var packet = {
             id : 'point',